fix conversion

+++ models.go
@@ -3,5 +3,5 @@ package dataversego
 type Authorization struct {
 	access_token  string
 	dataverse_url string
-	expires_in    int
+	expires_in    int64
 }
